Create Horizon-Builder-Requirements.md
kick starting the requirements doc. 
really considering whether to use the table structure or not, as maintaining it in .md is a bit of a pain. but it will be better overall if i can get it to work.... 
have a read and fill in your role. IDK what to call everyone. lol. 
I'll get filling with the reqs tomorrow when i have a minute.

Signed-off-by: Saerbhreathach <156859956+Saerbhreathach@users.noreply.github.com>